# Rework the commitQueue and commitPipeline syncing